LayoutTests/platform/mac-wk2/accessibility/mac/editable-page-lazy-spellchecking-expected.txt

This test ensures that the proper attributes are present when "lazy" spellchecking happens in a page marked as editable.

Attributed string with range: Attributes in range {0, 45}:
AXDidSpellCheck: 0
AXFont: {
    AXFontFamily = Times;
    AXFontName = "Times-Roman";
    AXFontSize = 16;
}
AXForegroundColor: (kCGColorSpaceICCBased; kCGColorSpaceModelRGB; sRGB IEC61966-2.1) headroom = 1.000000 ( 0 0 0 1 )
AXBackgroundColor: (kCGColorSpaceICCBased; kCGColorSpaceModelRGB; sRGB IEC61966-2.1) headroom = 1.000000 ( 0 0 0 0 )
wrods is misspelled aab lotsi nowadays. euep.

PASS successfullyParsed is true

TEST COMPLETE
wrods is misspelled aab lotsi nowadays. euep.
